bind click handler early
Currently, the click event handler for `.docs_poster--loader` is attached once the YouTube player is ready. This means interaction to the element is blocked until the js is downloaded, so early clicks are lost.

This changes the behaviour. Now, the click event handler is bound early and it invokes the download of the YouTube js. This means you can interact with the `.docs_poster--loader` element as soon as the page loads, rather than having to wait until the js is loaded.